Extract the palette, system-action, uninstall and custom-command coordinators
Completes the AppCore decomposition: 1021 -> 746 lines, four coordinators
joining the two from phase 24. PaletteWindowController keeps sole ownership
of the frame -- it is absent from this diff -- so the coordinator decides
which mode to show, never where or how big. DialogController stays
single-owned so a held hotkey still cannot stack dialogs. runWindowCommand
stays on AppCore; one method does not need a coordinator.

AuxWindowController and VolumeHUDController move ownership with the features
that use them; pickVolume needed a facade because DialogController may not.

AppCore is not yet under 300 lines and cannot be inside this phase: ~99 lines
are the palette types phase 28 extracts, ~99 are forwarders phase 32 deletes,
and ~195 lines of palette-UI actions have no owning phase at all. Recorded in
progress/25 as the gap phase 25b should close.
