+++ docs/refactor/ROADMAP.md
@@ -205,7 +205,7 @@ Update this table as phases land. `Blocked` requires a note in the phase's progr
 | 22  | Complete    | `refactor/22-screens-clipboard-and-calculator-history` | 2026-08-05 | **The calc card is a row, not an offset.** `rows` is `[.calc(result)] + entries.map(Row.entry)` and every lookup goes through `row(at:)`, so `calcCount` is absent from both new files — the pattern phase 23 reuses. The follow key still reads `store.items.first?.id`; its `vm.mode` guard is now structural (the handler only exists while the screen is mounted) and `scroll` is injected as a `scrollToFollow` closure. One new defaulted protocol member, `hasPrimaryAction(at:)` — without it the error card's "no pill, no ⌘K" rule regresses, since `calcResult` narrows to the launcher here. ⌘P/⌘⌫ route by `as?` downcast like phase 21; `delete` returns `Void` (the screen owns ⌘⌫ even with nothing to delete) while `pin` returns `Bool`. `RootPaletteView` 982 → 832 (−150); 7 files, +424/−292 against an expected 7 and +330/−290. `palette-selection-test` 111,066 → **111,108**, deliberate-break proof passed (exit 1). **`regression.md` waived** — AC4–AC7 and AC10 structural only, including the query-then-copy check. **Unblocks phase 23.** |
 | 23  | Complete    | `refactor/23-screen-launcher-and-offset-collapse` | 2026-08-05 | **M3 closes.** Every mode is a `PaletteScreen`; `calcCount` greps empty and `switch vm.mode` drops 3 → 1 — the survivor is the factory, now non-optional and exhaustive, which deletes the `if let screen` fork at every call site. `LauncherView.swift` is **absent from the diff**, so the nine-section table and its type annotation are byte-identical by construction. `selectionIsRunning` deliberately stays `@State` on the view: a screen has no storage, and reading `runningApps` from `actions(at:)` would put `RunningAppsMonitor` on the palette's body. ⌃⇧Q and ⌘1–⌘5 route by `as?` downcast — safe, not just equivalent, since `paletteIsCollapsed` already implies launcher mode. **AC2 not met** — 665 lines, not ~400: the doc's target assumed a 1126-line file, and the 534 that remain are all things the phase says stay, plus 128 lines of palette chrome **no phase has a home for** (a gap in 29, raised in `progress/23`). Diff +292/−199 against an expected +450/−800 — not net negative, because the file started at 832, not 1126. Debug **and Release** clean (37 s, annotation holds), 0 → 0 warnings, all 17 harnesses green, `palette-selection-test` 111,684. **`regression.md` not run** — waived, so AC4–AC10 are structural only, including the four ↑/↓ walks the risk register calls the real gate. **Unblocks M4.** |
 | 24  | Complete    | `refactor/24-quicklink-and-snippet-coordinators` | 2026-08-05 | **M4 opens.** The two largest orchestration blocks leave `AppCore`: 1342 → 1021 (**−321**, not −345 — 345 lines of body left, ~24 came back as the 13 forwarders and the 2 façade methods the phase itself requires). Both hard gates hold: `AppCore.shared` greps empty in both new files, `DialogController()` stays **1**. **AC3 was proven by direct body diff, not by `git diff -M`** — rename detection can't pair these bodies while `AppCore.swift` still exists, so the pre-change blocks were extracted from `HEAD` and diffed line-by-line; every difference is a receiver rename or `private func` → `func`. Both gating orders byte-identical. **The "+2 properties" budget is short by two**: `reportFailure` and `showMessage` were needed because `dialogs`/`messageHUD` are private and a coordinator may not own its own — phase 25 hits the same wall with `volumeHUD.show` and `dialogs.pickVolume`, neither of which it names. The two coordinators use **different injection shapes on purpose** — `SnippetExpansionCoordinator` holds no `AppCore` reference at all (one `showMessage` closure), `QuicklinkCoordinator` takes `unowned core` for six distinct capabilities. Debug **and** Release clean, 0 → 0 warnings, all 17 harnesses green including both phase gates. **`regression.md` waived** — so AC4–AC6 are structural only, including the "feature off, press the bound shortcut" funnel test the phase calls its single most valuable check. **Unblocks phase 25.** |
-| 25  | Not started |                                            |            |                                                                                                                                                                                                                                        |
+| 25  | Complete    | `refactor/25-remaining-coordinators`       | 2026-08-05 | **`AppCore` stops being the only orchestrator**: 1021 → 746 (−275), four coordinators, 2 → 6. All hard gates hold — `PaletteWindowController` is **absent from the diff** (so the frame, the session anchor and pop-to-root are byte-identical), `DialogController()` stays **1**, `AppCore.shared` greps empty in all four new files, and `runWindowCommand` stays put. Moves proven by direct body diff against `HEAD`, as in 24: every difference is a receiver rename or the one authorised `targetApp` helper, and the three pinned sequences (uninstall's ten steps, `perform`'s confirmation switch, `runCustomCommand`'s gate) are byte-identical. **`pickVolume` needed the third façade** `progress/24` predicted; `volumeHUD` and `auxWindows` did **not** — both moved ownership with their features instead. **AC2 not met — 746, not ~300, and the phase doc contradicts itself**: its own "net −370" budget lands at ~651. Of what remains, ~99 lines are the three palette types **phase 28 already extracts by name**, ~99 are forwarders **32** deletes, ~227 are composition root and things the phase says stay — and **~195 lines of palette-UI actions no phase owns**, which blocks 32's objective. Debug **and** Release clean, 0 → 0 warnings, all 17 harnesses green, `palette-selection-test` 111,684. **`regression.md` waived** — AC5–AC10 structural only, including the compact top-edge and held-hotkey checks the kickoff names. **Unblocks 26 and 32.** M4's DoD is **not** yet true; **25b** is proposed to close it. |
 | 26  | Not started |                                            |            |                                                                                                                                                                                                                                        |
 | 27  | Not started |                                            |            |                                                                                                                                                                                                                                        |
 | 28  | Not started |                                            |            |                                                                                                                                                                                                                                        |
